requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
ccxt==4.5.34
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coincurve==21.0.0
cryptography==46.0.3
curl_cffi==0.13.0
deprecation==2.1.0
exceptiongroup==1.3.1
feedparser==6.0.12
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
multitasking==0.0.12
narwhals==2.15.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.2
postgrest==2.27.2
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.27.2
referencing==0.36.2
requests==2.32.5
rich==14.3.1
rpds-py==0.27.1
rsa==4.9.1
schedule==1.2.2
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8.3
storage3==2.27.2
streamlit==1.50.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
yfinance==1.1.0
